Removed "expected_max_run_time_ms", Added "md5sum" body check.
This checkin removes the "expected_max_run_time_ms". This was used by one of our tests only and I think it need to be rethought what this parameter really needs to do. And that is a timeout for a request. We should implement that rather than not timinig out and deciding later that it took too long.

Also, removed the direct translation of a binary file into a response body. This was a limited choice and we could not check such a binary body together with a status code other than 200. Now, we don't support direct binary comparison any more, instead we generate an "md5sum" and a pesudo JSON body for binary bodies. The template helper function "md5sum" can be used to write the expected response for a request.