template function "file": pass all parameters
This code simplifies the implementation of the template function "file".

Now, we allow all parameters to be passed into the template render. Since we are using a Map now which is passed into the template, ".Something" will not cause an render error in the template any more as the map is accessed using the key "Something" and, assumed it doesnt exist, return the nil interface. Before, this was caused a template rendering error.

With this change a call to a "file" can implement to work with a flexibel amount of parameters which wasn't possible before.