template_funcs: changed "file", added "file_render", removed "parse_csv"
This patch addresses a performance problem in file loading function "file". Previously we would render the data as Go template resulting in bad performance for huge files.

Since most of our usages are not expecting file to be a template, renamed "file" to "file_render". "file" now only reads the file with no further processing.

While at it removed undocumented "parse_csv" (unused by us).

Also changed "file_csv" to default to delimiter ',' if none is given. This makes it easier to use this function.

Fixed tests which compare error strings. This is a no go, cause error depend on stdlib as well as 3rd party. So unless the error has a name, never compare the error strings...

see #63690