Fixed ":control" matches in Arrays
If a :control was used inside an Array to match data, it only worked properly if only one element was checked in the Array or the checked item would not be an nested more complex Object.

A matching like this would not fail even if the "url" started with "http".

[
    {
        "_id": 1,
        "versions": {
            "preview": {
                "url:control": {
                    "match": "^henk"
                }
            }
        }
    },
    {
        "_id": 2
    },
    {
        "_id": 3
    }
]

This patch fixed also the "is_number" comparison which showed buggy after the first fix.